--- src/components/sections/index.js
@@ -1,3 +1,0 @@
-import LandingTop from './landing_top'
-
-export { LandingTop }
