+++ src/pages/home/sections/index.js
@@ -1,0 +1,5 @@
+import LandingTop from './landing_top';
+import LandingMiddle from './landing_middle'
+import LandingBottom from './landing_bottom'
+
+export {LandingTop, LandingMiddle, LandingBottom}
