+++ src/pages/home/sections/slideshow.js
@@ -6,7 +6,7 @@ import Paper from '@material-ui/core/Paper';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 import Container from '@material-ui/core/Container';
-import items from '../../data/technologies.js'
+import items from '../../../data/technologies.js'
 
 
 // importing images
